requirements.txt
asgiref==3.3.1
astroid==2.5
Django==3.1.6
djangorestframework==3.12.2
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mysql==0.0.2
mysqlclient==2.0.3
p==0.0.0
pylint==2.7.1
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
